Step-2

+++ src/controllers/cards.ts
@@ -1,7 +1,9 @@
-import { Request, Response, NextFunction } from "express";
-import { Card } from "../models/card";
-import { successResponse } from "../helpers";
-import { NotFoundError } from "../types/errors";
+import { Request, Response, NextFunction } from 'express';
+import { POST_NOT_FOUND_MESSAGE } from '../types/errors';
+import { Card } from '../models/card';
+import { successResponse } from '../helpers';
+import NotFoundError from '../types/Errors/NotFoundError';
+import ForbiddenError from '../types/Errors/ForbiddenError';
 
 const getCards = (req: Request, res: Response, next: NextFunction) => {
   Card.find({})
@@ -16,14 +18,20 @@ const createCard = (req: Request, res: Response, next: NextFunction) => {
 };
 
 const deleteCard = (req: Request, res: Response, next: NextFunction) => {
-  Card.deleteOne({ _id: req.params.cardId })
-    .then((data) => {
-      if (data.deletedCount === 0) {
-        throw new NotFoundError("Пост не найден.");
-      }
-      res.status(200).send(successResponse({ message: "Пост удалён" }));
-    })
-    .catch(next);
+  Card.findById(req.params.cardId).then((card) => {
+    if ((card?.owner as any).equals((req as any).user._id)) {
+      Card.deleteOne({ _id: req.params.cardId })
+        .then((data) => {
+          if (data.deletedCount === 0) {
+            throw new NotFoundError(POST_NOT_FOUND_MESSAGE);
+          }
+          res.status(200).send(successResponse({ message: 'Пост удалён' }));
+        })
+        .catch(next);
+    } else {
+      throw new ForbiddenError('Пользователь не может удалить чужую карточку.');
+    }
+  }).catch(next);
 };
 
 const likeCard = (req: Request, res: Response, next: NextFunction) => {
@@ -37,7 +45,12 @@ const likeCard = (req: Request, res: Response, next: NextFunction) => {
       runValidators: true,
     },
   )
-    .then((data) => res.status(200).send(successResponse(data)))
+    .then((data) => {
+      if (!data) {
+        throw new NotFoundError(POST_NOT_FOUND_MESSAGE);
+      }
+      res.status(200).send(successResponse(data));
+    })
     .catch(next);
 };
 
@@ -52,7 +65,12 @@ const unlikeCard = (req: Request, res: Response, next: NextFunction) => {
       runValidators: true,
     },
   )
-    .then((data) => res.status(200).send(successResponse(data)))
+    .then((data) => {
+      if (!data) {
+        throw new NotFoundError(POST_NOT_FOUND_MESSAGE);
+      }
+      res.status(200).send(successResponse(data));
+    })
     .catch(next);
 };
